refactor(server-oauth2-mock): move portal-discovery, file-user-store, and ensurePortInUrl into seedwork
- discoverPortalConfigs + PortalOidcConfig moved to @cellix/server-oauth2-mock-seedwork
- createFileUserStore moved to @cellix/server-oauth2-mock-seedwork
- ensurePortInUrl merged into seedwork utils
- 22 portal-discovery tests, 2 file-user-store tests, 5 ensurePortInUrl tests moved to seedwork
- @apps/server-oauth2-mock reduced to app-specific wiring (index.ts + setup-environment.ts)
- seedwork manifest.md and README.md updated
