+++ packages/cellix/server-oauth2-mock-seedwork/src/index.ts
@@ -7,5 +7,9 @@ export { createMockOAuth2Manager } from './manager.ts';
 export { buildOidcRouter } from './router.ts';
 export { startMockOAuth2Server } from './server.ts';
 
+export { discoverPortalConfigs } from './portal-discovery.ts';
+export type { PortalOidcConfig } from './portal-discovery.ts';
+export { createFileUserStore } from './file-user-store.ts';
+
 export type { MockOAuth2Manager, MockOAuth2PortalConfig, MockOAuth2Registration, MockOAuth2ServerConfig, MockOAuth2ServerHandle, MockOAuth2User, MockOAuth2UserProfile, MockOAuth2UserStore } from './types.ts';
-export { AUTH_CODE_PREFIX, normalizeBaseUrl, normalizeOrigin, normalizeUrl, SAFE_NAME_RE } from './utils.ts';
+export { AUTH_CODE_PREFIX, ensurePortInUrl, normalizeBaseUrl, normalizeOrigin, normalizeUrl, SAFE_NAME_RE } from './utils.ts';
