+++ packages/cellix/server-oauth2-mock-seedwork/tests/file-user-store.test.ts
@@ -2,7 +2,7 @@ import * as fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
-import { createFileUserStore } from '../src/users.ts';
+import { createFileUserStore } from '../src/index.ts';
 
 function makeTempDir() {
 	return fs.mkdtempSync(path.join(os.tmpdir(), 'server-oauth2-mock-users-'));
