+++ packages/cellix/server-oauth2-mock-seedwork/tests/portal-discovery.test.ts
@@ -2,7 +2,7 @@ import * as fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { discoverPortalConfigs } from '../src/portal-discovery.ts';
+import { discoverPortalConfigs } from '../src/index.ts';
 
 function makeTempAppsDir() {
 	return fs.mkdtempSync(path.join(os.tmpdir(), 'server-oauth2-mock-tests-'));
